Workaround out-of-sync 1 PPS signal in Evans Hall
The code was written assuming that the 1 PPS signal and the host
computer's NTP time are fairly closely aligned (actually, it was fairly
lenient).  In the basement of Evans Hall we (still) have no GPS antenna
for the GPS clock that generates our 1 PPS and 10 MHz reference.  The
"free-running" 1 PPS signal has drifted far enough away from the NTP
top-of-second that the code did not sync multiple F engines to the same
1 PPS pulse (the arm commands went out right around a 1 PPS instead of
just after it).

To work around this, the code now reads the first F engine's sync_count
register in a tight loop until it increments.  At that point it assumes
that all F engines have just received a 1 PPS pulse so it is safe to
send the arm command out to all F engines before the next 1 PPS pulse.