More F engine initialization functionality
paper_feng_init.rb now sets the fft_shift register, the noise source
seed registers, and the EQ coefficients.  The fft_shift and EQ values
can (optionally) be specified on the command line.  One EQ value can be
given and that one value is used for all channels of all inputs.  See
the diff or the command line help for the default fft_shift and EQ
values.

If syncing is done (the defaut), then the seed registers are initialized
and the noise generators sync'd.  The seed value used is 0x11111111.  At
this time, the seed value cannot be specified and all four seed
registers are set to the same value.

If syncing is NOT done (i.e the -s or --no-sync option is passed), the
seed registers are not updated and the noise sources are not sync'd.