Overhaul paper_redis_spec.rb
Now it generates a PNG file suitable for sending to the paper blog and
outputs the body of a message that can be piped into "mail -~".  The
message uses "tilde escapes" to set the subject line and attach the
plot, so the "-~" option must be passed to mail to enable processing of
these escape sequences.